requirements.txt
attrs>=23.1.0
chemistry-tools[formulae]>=v1.0.0b2
domdf-python-tools>=3.6.1
pymassspec>=2.3.0
